pyproject.toml
[project]
name = "korean-bus-terminal"
version = "0.1.0"
description = ""
authors = [
    {name = "dellyu03",email = "seunghwan081@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
